requirements.txt
# instructions 
# conda create -n cvpr23 python=3.9.12
# conda install pytorch==1.12.0 torchvision==0.13.0 torchaudio==0.12.0 cudatoolkit=11.3 -c pytorch
# pip install -r requirements.txt


pytorch-lightning==1.5.2
einops==0.3.2
opencv-python==4.5.4.58
matplotlib==3.5.0
imageio==2.10.4
scipy==1.8.0
imageio-ffmpeg==0.4.5
open3d==0.14.1
kornia==0.4.1
pandas==1.4.3
jupyter
tqdm
torch_optimizer
wandb
plotly

# visualization
# PyMCubes
# pycollada
# trimesh
# pyglet
# networkx
# plyfile
# open3d